Now outputting data to cassandra

+++ src/main/java/com/gaddieind/hadoop/ExceptionMapper.java
@@ -2,6 +2,7 @@ package com.gaddieind.hadoop;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -11,7 +12,7 @@ import java.io.IOException;
 /**
  * Created by agaddie on 10/3/2015.
  */
-public class ExceptionMapper extends Mapper<LongWritable, Text, Text, DoubleWritable> {
+public class ExceptionMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
 
     @Override
     protected void map(LongWritable key, Text value, Context context) throws IOException, InterruptedException {
@@ -21,7 +22,7 @@ public class ExceptionMapper extends Mapper<LongWritable, Text, Text, DoubleWrit
 
         if(startOfFailingClass > 0 && endOfFailingClass > 0) {
             String failingClassInfo = stacktrace.substring(startOfFailingClass + 1, endOfFailingClass);
-            context.write(new Text(failingClassInfo), new DoubleWritable(1));
+            context.write(new Text(failingClassInfo), new IntWritable(1));
         }
     }
 }
